refactor(#3218): move DisabledHomepage to apps/secret/views
Per project convention, each app under src/apps/ mirrors the top-level
src layout — so views/ is the per-app analog of src/views/. The
existing AccessDenied.vue was misshelved in conceal/ (which is for
secret-creation forms — incoming, homepage); moving the file to its
correct home as part of this refactor.

Co-locates the disabled/ subtree (composable, variants, README) with
the dispatcher under views/, and mirrors the test path
(src/tests/apps/secret/{conceal => views}/).

Updates:
- src/apps/secret/conceal/Homepage.vue: import path
- src/views/PreviewDisabled.vue: import path
- src/tests/apps/secret/views/useDisabledConfig.spec.ts: import path
- inline file-path comments + README directory diagram + test path
